Integrate Harvard Art Museum API + Decks
The getUniqueImageInRound and getImageForPlayers endpoints now take into account whether or not the selected deck is for the Harvard Art Museum. Requests use Mayukh's API key and the imageid in the entries to uniquely retrieve images from the API.

The decks endpoint has now been repaired so that it takes in parameters instead of the previously mistaken form data.