Modifications to checkEmailValidationCode Endpoint
The endpoint now takes the user_uid and verification code as a json payload now. With the user_uid, the search in the user table will be narrowed down to a single row where the provided code will be compared to the payload, and a response will be returned depending on their similarity. A couple edge cases were taken care of as well.